Move card logic to the card itself
